+++ code/modelGNN_updates.py
@@ -30,41 +30,41 @@ class ReGGNN(nn.Module):
         if self.att_op == att_op_dict['concat']:
             self.out_dim = hidden_size * 2
 
-        self.emb_encode = nn.Linear(feature_dim_size, hidden_size).double()
+        self.emb_encode = nn.Linear(feature_dim_size, hidden_size).float()
         self.dropout_encode = nn.Dropout(dropout)
-        self.z0 = nn.Linear(hidden_size, hidden_size).double()
-        self.z1 = nn.Linear(hidden_size, hidden_size).double()
-        self.r0 = nn.Linear(hidden_size, hidden_size).double()
-        self.r1 = nn.Linear(hidden_size, hidden_size).double()
-        self.h0 = nn.Linear(hidden_size, hidden_size).double()
-        self.h1 = nn.Linear(hidden_size, hidden_size).double()
-        self.soft_att = nn.Linear(hidden_size, 1).double()
-        self.ln = nn.Linear(hidden_size, hidden_size).double()
+        self.z0 = nn.Linear(hidden_size, hidden_size).float()
+        self.z1 = nn.Linear(hidden_size, hidden_size).float()
+        self.r0 = nn.Linear(hidden_size, hidden_size).float()
+        self.r1 = nn.Linear(hidden_size, hidden_size).float()
+        self.h0 = nn.Linear(hidden_size, hidden_size).float()
+        self.h1 = nn.Linear(hidden_size, hidden_size).float()
+        self.soft_att = nn.Linear(hidden_size, 1).float()
+        self.ln = nn.Linear(hidden_size, hidden_size).float()
         self.act = act
 
     def gatedGNN(self, x, adj):
         a = torch.matmul(adj, x)
         # update gate
-        z0 = self.z0(a.double())
-        z1 = self.z1(x.double())
+        z0 = self.z0(a.float())
+        z1 = self.z1(x.float())
         z = torch.sigmoid(z0 + z1)
         # reset gate
-        r = torch.sigmoid(self.r0(a.double()) + self.r1(x.double()))
+        r = torch.sigmoid(self.r0(a.float()) + self.r1(x.float()))
         # update embeddings
-        h = self.act(self.h0(a.double()) + self.h1(r.double() * x.double()))
+        h = self.act(self.h0(a.float()) + self.h1(r.float() * x.float()))
 
         return h * z + x * (1 - z)
 
     def forward(self, inputs, adj, mask):
         x = inputs
         x = self.dropout_encode(x)
-        x = self.emb_encode(x.double())
+        x = self.emb_encode(x.float())
         x = x * mask
         for idx_layer in range(self.num_GNN_layers):
             if self.residual:
-                x = x + self.gatedGNN(x.double(), adj.double()) * mask.double()  # add residual connection, can use a weighted sum
+                x = x + self.gatedGNN(x.float(), adj.float()) * mask.float()  # add residual connection, can use a weighted sum
             else:
-                x = self.gatedGNN(x.double(), adj.double()) * mask.double()
+                x = self.gatedGNN(x.float(), adj.float()) * mask.float()
         # soft attention
         soft_att = torch.sigmoid(self.soft_att(x))
         x = self.act(self.ln(x))
@@ -101,20 +101,20 @@ class ReGCN(nn.Module):
                 self.gnnlayers.append(RelationalGraphConvLayer(feature_dim_size, hidden_size, num_bases, num_relations, dropout, bias=True))  # bias=False
             else:
                 self.gnnlayers.append(RelationalGraphConvLayer(hidden_size, hidden_size, num_bases, num_relations, dropout, bias=True))
-        self.soft_att0 = nn.Linear(32, 1).double()
-        self.ln0 = nn.Linear(1, 32).double()
+        self.soft_att0 = nn.Linear(32, 1).float()
+        self.ln0 = nn.Linear(1, 32).float()
 
-        self.q = nn.Linear(hidden_size, hidden_size).double()
-        self.k = nn.Linear(hidden_size, hidden_size).double()
-        self.v = nn.Linear(hidden_size, hidden_size).double()
-        self.soft_att = nn.Linear(hidden_size, 1).double()
-        self.ln = nn.Linear(hidden_size, hidden_size).double()
+        self.q = nn.Linear(hidden_size, hidden_size).float()
+        self.k = nn.Linear(hidden_size, hidden_size).float()
+        self.v = nn.Linear(hidden_size, hidden_size).float()
+        self.soft_att = nn.Linear(hidden_size, 1).float()
+        self.ln = nn.Linear(hidden_size, hidden_size).float()
         self.act = act
        
     def forward(self, inputs, adj, mask):
         x = inputs
         x = x.permute(0, 1, 3, 2)
-        attn_token = torch.tanh(self.soft_att0(x.double()).double())
+        attn_token = torch.tanh(self.soft_att0(x.float()).float())
         attn_token = torch.nn.functional.softmax(self.ln0(attn_token), dim=3)
         x = attn_token * x
         x = torch.sum(x, 3)
@@ -128,9 +128,9 @@ class ReGCN(nn.Module):
                 else:
                     x = self.gnnlayers[idx_layer](adj, x) * mask
         # soft attention
-        query = self.q(x.double()).double()
-        key = self.k(x.double()).double()
-        value = self.v(x.double()).double()
+        query = self.q(x.float()).float()
+        key = self.k(x.float()).float()
+        value = self.v(x.float()).float()
         attn_node = torch.matmul(query, key.permute(0, 2, 1))
         attn_node = attn_node / math.sqrt(self.hidden_size)
         attn_probs = torch.nn.functional.softmax(attn_node, dim=-1)
@@ -145,38 +145,38 @@ class GGGNN(nn.Module):
     def __init__(self, feature_dim_size, hidden_size, num_GNN_layers, dropout, act=nn.functional.relu):
         super(GGGNN, self).__init__()
         self.num_GNN_layers = num_GNN_layers
-        self.emb_encode = nn.Linear(feature_dim_size, hidden_size).double()
+        self.emb_encode = nn.Linear(feature_dim_size, hidden_size).float()
         self.dropout_encode = nn.Dropout(dropout)
-        self.z0 = nn.Linear(hidden_size, hidden_size).double()
-        self.z1 = nn.Linear(hidden_size, hidden_size).double()
-        self.r0 = nn.Linear(hidden_size, hidden_size).double()
-        self.r1 = nn.Linear(hidden_size, hidden_size).double()
-        self.h0 = nn.Linear(hidden_size, hidden_size).double()
-        self.h1 = nn.Linear(hidden_size, hidden_size).double()
-        self.soft_att = nn.Linear(hidden_size, 1).double()
-        self.ln = nn.Linear(hidden_size, hidden_size).double()
+        self.z0 = nn.Linear(hidden_size, hidden_size).float()
+        self.z1 = nn.Linear(hidden_size, hidden_size).float()
+        self.r0 = nn.Linear(hidden_size, hidden_size).float()
+        self.r1 = nn.Linear(hidden_size, hidden_size).float()
+        self.h0 = nn.Linear(hidden_size, hidden_size).float()
+        self.h1 = nn.Linear(hidden_size, hidden_size).float()
+        self.soft_att = nn.Linear(hidden_size, 1).float()
+        self.ln = nn.Linear(hidden_size, hidden_size).float()
         self.act = act
 
     def gatedGNN(self, x, adj):
         a = torch.matmul(adj, x)
         # update gate
-        z0 = self.z0(a.double())
-        z1 = self.z1(x.double())
+        z0 = self.z0(a.float())
+        z1 = self.z1(x.float())
         z = torch.sigmoid(z0 + z1)
         # reset gate
-        r = torch.sigmoid(self.r0(a.double()) + self.r1(x.double()))
+        r = torch.sigmoid(self.r0(a.float()) + self.r1(x.float()))
         # update embeddings
-        h = self.act(self.h0(a.double()) + self.h1(r.double() * x.double()))
+        h = self.act(self.h0(a.float()) + self.h1(r.float() * x.float()))
 
         return h * z + x * (1 - z)
 
     def forward(self, inputs, adj, mask):
         x = inputs
         x = self.dropout_encode(x)
-        x = self.emb_encode(x.double())
+        x = self.emb_encode(x.float())
         x = x * mask
         for idx_layer in range(self.num_GNN_layers):
-            x = self.gatedGNN(x.double(), adj.double()) * mask.double()
+            x = self.gatedGNN(x.float(), adj.float()) * mask.float()
         return x
 
 
@@ -202,8 +202,8 @@ class GraphConvolution(torch.nn.Module):
 
     def forward(self, input, adj):
         x = self.dropout(input)
-        support = torch.matmul(x.double(), self.weight.double())
-        output = torch.matmul(adj.double(), support.double())
+        support = torch.matmul(x.float(), self.weight.float())
+        output = torch.matmul(adj.float(), support.float())
         if self.bias is not None:
             output = output + self.bias
         return self.act(output)
@@ -273,10 +273,10 @@ class RelationalGraphConvLayer(Module):
         supports = []
         A = A.permute(1, 0, 2, 3)
         for i in range(self.num_rel):
-            supports.append(torch.matmul(A[i].double(), X.double()))
+            supports.append(torch.matmul(A[i].float(), X.float()))
 
         tmp = torch.cat(supports, dim=2)
-        out = torch.matmul(tmp.double(), weights.double())  # shape(#node, output_size)
+        out = torch.matmul(tmp.float(), weights.float())  # shape(#node, output_size)
 
         if self.bias is not None:
             out += self.bias.unsqueeze(0)
